The 0-tick Patch
@ High level:
* onVarClientIntChanged used to detect inventory
  swaps simply perform the attack again if we are
  early enough in the tick since that means the
  user has hit with the swapped weapon.
* track spell book, track spell sounds, track
  projectiles aimed at our target.
* Change "attack type" code completely, use above
  tracking to help better detect powered staves.
* The salamander fix (need a spot anim tweak to
  detect those attacks).

@ Bug fixes:
* Blowpipe no longer incidentally triggering after
  stopping attacking (FIXED)
* Manual casting detection
* No longer rendering negative numbers should the
  plugin get confused
* Add dawnbringer (ADDED)
* Harm support (ADDED)

@ Implementation details

New store of PoweredStaves their weapon IDs and
the projectiles and animations they have, this
already helps because animations know what spell
book they come from, so we can cut a lot of cases
out by being on thralls (stay winning).

onVarClientIntChanged is the key to getting
0-ticking working, without this varclient we are
locked to the onGameTick which will have the
equipped weapon changed out from underneath it if
they swap weapons in that tick.

Sounds, spell books, and animations all help make
powered staves better. But there's still edge
cases, the no animation when barraging after
blowpiping is super annoying to fix. It actually
might not be fixable with the current APIs.

Finally I think this deserves a version bump
:smile:

+++ dev.ps1
@@ -1,20 +1,27 @@
-$devlocation = "C:\Users\Lexer\Repos\runelite\runelite-client\src\main\java\net\runelite\client\plugins\improvedattacktimer"
+.\gradlew.bat build
+if (-not $?) { exit $? }
+.\gradlew.bat test
+if (-not $?) { exit $? }
+$repo = Get-Location
+$devlocation = "$repo\..\runelite\runelite-client\src\main\java\net\runelite\client\plugins\improvedattacktimer"
 
 $oldDevFiles = Get-ChildItem -Path $devlocation -Filter *.java -Recurse -File
 foreach ($oldFile in $oldDevFiles) {
-    echo "Removing $devlocation\$oldFile"
+    Write-Output "Removing $devlocation\$oldFile"
     Remove-Item -Path "$devlocation\$oldFile"
 }
 
-$srcFilePath = "C:\Users\Lexer\Repos\improvedattacktimer\src\main\java\com\improvedattacktimer"
+$srcFilePath = "$repo\src\main\java\com\improvedattacktimer"
 $srcFiles = Get-ChildItem -Path $srcFilePath -Filter *.java -Recurse -File
 $oldPackage = "com.improvedattacktimer"
 $newPackage = "net.runelite.client.plugins.improvedattacktimer"
+$debugString = "LOCAL_DEBUGGING_20ad04d7 = false;" # should be committed to the repo off.
+$debugOnString = "LOCAL_DEBUGGING_20ad04d7 = true;"
 foreach ($srcFile in $srcFiles) {
-    echo "Copying file $srcFile -> $devlocation"
+    Write-Output "Copying file $srcFile -> $devlocation"
     Copy-Item -Path "$srcFilePath\$srcFile" "$devlocation\$srcFile"
-    echo "Updating package $oldPackage -> $newPackage | FILE: $devlocation\$srcFile"
-    (Get-Content "$devlocation\$srcFile").Replace($oldPackage, $newPackage ) | Set-Content "$devlocation\$srcFile"
+    Write-Output "Updating package $oldPackage -> $newPackage | FILE: $devlocation\$srcFile"
+    (Get-Content "$devlocation\$srcFile").Replace($oldPackage, $newPackage).Replace($debugString, $debugOnString) | Set-Content "$devlocation\$srcFile"
 }
-echo "Done :)"
+Write-Output "Done :)"
 exit 0
